Fixed frame spacing between dynamic beats.
Dynamic beats now spawn with a spacing equal to the bpm of the song divided by the number of beats in the beatpattern object.

In the case that this is not a whole number, the fractional part of the number is saved and then added to the frame spacing whenever it exceeds 1.

This works better than expected however it still may result in out of sync beats.